Add files via upload
This project focuses on developing a machine learning model to classify Iris flowers into three species: setosa, versicolor, and virginica. The classification is based on the measurements of their sepals and petals. Using the well-known Iris dataset, which includes these measurements categorized by species, the goal is to train a model that accurately predicts the species of Iris flowers. This dataset is widely utilized for introductory classification tasks, making it an ideal choice for demonstrating basic machine learning concepts.