IRIS FLOWER CLASSIFICATION
Iris Flower Classification
This project leverages the Iris flower dataset, which consists of measurements of three species of Iris flowers: setosa, versicolor, and virginica. These species can be distinguished based on the measurements of their sepals and petals. The objective of this project is to develop a machine learning model that can accurately classify Iris flowers into their respective species based on these measurements.

Using the Iris dataset, which is widely used for introductory classification tasks, we train and evaluate various machine learning algorithms to achieve accurate classification of the Iris flower species.